remove inbox.logging.sentry
