+++ inbox/mailsync/service.py
@@ -9,10 +9,10 @@ from sqlalchemy.exc import OperationalError
 
 from inbox.config import config
 from inbox.contacts.remote_sync import ContactSync
+from inbox.error_handling import log_uncaught_errors
 from inbox.events.remote_sync import EventSync, GoogleEventSync
 from inbox.heartbeat.status import clear_heartbeat_status
 from inbox.logging import get_logger
-from inbox.logging.sentry import log_uncaught_errors
 from inbox.mailsync.backends import module_registry
 from inbox.models import Account
 from inbox.models.session import global_session_scope, session_scope
