+++ inbox/models/account.py
@@ -20,8 +20,8 @@ from sqlalchemy.orm.session import Session
 from sqlalchemy.sql.expression import false
 
 from inbox.config import config
+from inbox.error_handling import log_uncaught_errors
 from inbox.logging import get_logger
-from inbox.logging.sentry import log_uncaught_errors
 from inbox.models.base import MailSyncBase
 from inbox.models.calendar import Calendar
 from inbox.models.mixins import (
