+++ inbox/transactions/actions.py
@@ -38,10 +38,10 @@ from inbox.actions.base import (
 )
 from inbox.config import config
 from inbox.crispin import writable_connection_pool
+from inbox.error_handling import log_uncaught_errors
 from inbox.events.actions.base import create_event, delete_event, update_event
 from inbox.ignition import engine_manager
 from inbox.logging import get_logger
-from inbox.logging.sentry import log_uncaught_errors
 from inbox.models import ActionLog, Event
 from inbox.models.session import session_scope, session_scope_by_shard_id
 from inbox.util.concurrency import retry_with_logging
